refactor: fix forms anti-pattern and improve component naming
- Remove nested useForm from GameSettingsForm
- Rename GameSettingsForm to Difficulty
- Make Difficulty a simple controlled select component
- Apply classNames utility to Button, StartForm, Difficulty
- Add comprehensive JSDoc documentation to components
- Rename CSS modules to match new component names
