+++ src/components/index.jsx
@@ -1,6 +1,5 @@
 ﻿export { default as Button } from './ui/Button';
-export { default as PlayerNameForm } from './ui/PlayerNameForm';
-export { default as GameSettingsForm } from './ui/GameSettingsForm';
+export { default as Difficulty } from './ui/Difficulty';
 export { default as StartForm } from './ui/StartForm';
 export { default as Player } from './ui/Player';
 export { default as ScoresList } from './ui/ScoresList';
