+++ src/components/ui/Button.jsx
@@ -1,5 +1,19 @@
+import { classNames } from "../../utils/classNames.js";
 import { Button as styles } from "../../css";
 
+/**
+ * Button Component
+ * 
+ * Reusable button with multiple variants and sizes.
+ * 
+ * @param {React.ReactNode} children - Button content
+ * @param {string} variant - Button style variant (primary, secondary, success, danger)
+ * @param {string} size - Button size (small, medium, large)
+ * @param {boolean} disabled - Whether button is disabled
+ * @param {string} type - Button type (button, submit, reset)
+ * @param {string} className - Additional CSS classes
+ * @param {Function} onClick - Click handler
+ */
 const Button = ({
     children,
     variant = "primary",
@@ -10,15 +24,12 @@ const Button = ({
     onClick,
     ...props
 }) => {
-    const variantClass = styles[variant] || styles.primary;
-    const sizeClass = styles[size] || styles.medium;
-
-    const buttonClass = [
+    const buttonClass = classNames(
         styles.button,
-        variantClass,
-        sizeClass,
+        styles[variant],
+        styles[size],
         className
-    ].filter(Boolean).join(" ");
+    );
 
     return (
         <button
